Exclude HasItems from HasItemsAndComponents so as to let the components support DataView
